@@ -11,7 +11,7 @@
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <jersey.version>3.0.1</jersey.version>
+    <jersey.version>3.1.3</jersey.version>
     <jena.version>5.2.0</jena.version>
     <owlapi.version>5.5.1</owlapi.version>
     <hermit.version>1.4.5.519</hermit.version>
@@ -21,7 +21,7 @@
   </properties>
 
   <build>
-    <sourceDirectory>src</sourceDirectory>
+  <sourceDirectory>src</sourceDirectory>
     <resources>
       <resource>
         <directory>config</directory>
@@ -30,15 +30,19 @@
     <plugins>
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.7.0</version>
+        <version>3.11.0</version>
         <configuration>
-          <release>10</release>
+          <release>17</release>
         </configuration>
       </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-war-plugin</artifactId>
         <version>3.3.0</version>
+        <configuration>
+          <!-- Use WebContent folder as the webapp source so WEB-INF/web.xml is packaged -->
+          <warSourceDirectory>${project.basedir}/WebContent</warSourceDirectory>
+        </configuration>
       </plugin>
     </plugins>
   </build>
@@ -102,6 +106,12 @@
       <artifactId>jersey-media-jaxb</artifactId>
       <version>${jersey.version}</version>
     </dependency>
+    <!-- JAXB runtime (Jakarta XML Bind) required by Jersey's JAXB providers -->
+    <dependency>
+      <groupId>org.glassfish.jaxb</groupId>
+      <artifactId>jaxb-runtime</artifactId>
+      <version>4.0.3</version>
+    </dependency>
     <dependency>
       <groupId>org.glassfish.jersey.media</groupId>
       <artifactId>jersey-media-multipart</artifactId>
@@ -137,23 +147,18 @@
       <version>${log4j.version}</version>
     </dependency>
 
-    <!-- Servlet & JSTL -->
+    <!-- Jakarta Servlet & JSTL -->
     <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <version>4.0.1</version>
-      <scope>provided</scope>
+        <groupId>jakarta.servlet</groupId>
+        <artifactId>jakarta.servlet-api</artifactId>
+        <version>5.0.0</version>
+        <scope>provided</scope>
     </dependency>
     <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>jstl</artifactId>
-      <version>1.2</version>
-    </dependency>
-    <dependency>
-      <groupId>javax</groupId>
-      <artifactId>javaee-web-api</artifactId>
-      <version>8.0</version>
-      <scope>provided</scope>
+        <groupId>jakarta.platform</groupId>
+        <artifactId>jakarta.jakartaee-api</artifactId>
+        <version>10.0.0</version>
+        <scope>provided</scope>
     </dependency>
 
     <!-- Apache SIS EPSG -->
